requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
pathlib2>=2.3.7
plotly>=5.15.0
numpy>=1.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
mysql-connector-python>=8.0.0
openai-whisper>=20231117
librosa>=0.10.0
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0 